completed shh functionality for empty/non-empty house and temp monitoring

+++ Phase2/SmartHomeSimulator/src/UI/EditSimulation/EditSimulationController.java
@@ -64,7 +64,7 @@ public class EditSimulationController implements Initializable {
 
         menu.setDate(datePicker.getValue().toString());
         menu.setTime(newTime.getText());
-        menu.setTemp(Integer.parseInt(newTemp.getText()));
+        menu.setTemp(Float.parseFloat(newTemp.getText()));
         menu.setRoom(changeRoomBox.getValue());
         boolean userLeft = false;
